Basic Imagebox layer
Completely uncompiled so far. I copied CoordLayer and gutted it ...
surprisingly, still a lot of boilerplate code. Added in my params I
need. This is a good starting point for someone to make a new layer.

Next steps: add to Makefile, add to list of layers, try to compile, and
then, add drawing code! Then make a little tool, and an function to draw
images and I'm done! And then comes support for GPS crosshairs, etc... ok,
almost there...